cognito-fastapi/backend/requirements.txt
fastapi
uvicorn
python-jose
requests
boto3
python-dotenv
